Issue #339
Split Unit Tests into mutliple files

split cog_test.py into multiple files

turned tests dir into module by adding __init__.py
